Finished products page
but in terms of its content probably needs to be more like the idea of some prebuilt pillows that are then customized

NOTE: Not done the customization page. Need to get that started eventually.

The process of taking them from their prebuilt-->customization-->submit design in-->go to account and then to order